Refactoring renaming + ci setup :rocket:
