+++ scripts/test-ci.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env zsh
+
+# Build script for CI
+# install gatsbyjs cli and build
+
+npm install -g gatsby-cli
+npm install
+npm run test
