+++ src/logic/seo.ts
@@ -1,0 +1,5 @@
+export enum OgPageType {
+  article = "article",
+  website = "website",
+  profile = "profile",
+}
