--- src/logic/tag.ts
@@ -1,2 +1,0 @@
-export const generateTagLink = (tag: string) =>
-  `/blog/tags/${tag!.split(" ").join("-")}/`;
