+++ src/logic/url.ts
@@ -1,5 +1,8 @@
 import { WindowLocation } from "@reach/router";
 
+export const generateTagUrl = (tag: string) =>
+  `/blog/tags/${tag!.split(" ").join("-")}/`;
+
 export interface CurrentLocation {
   url: string;
   pathname: string;
